snakeberry/requirements.txt
pymongo[srv]==4.4.1
fastapi==0.101.1
pydantic==2.2.1
uvicorn==0.23.2
motor==3.2.0
beanie==1.21.0
python-dotenv==1.0.0
httpx==0.25.0